controller.js
Add logic so that the rm, rb only toggle on word change if not already active.

Add additional logic in parsing in the case that the entry does not include '__text' to just grab the dt value instead.
--

style.css

Add direction divs above the right and in-between to instruct the user on each step of the dictionary process.  Adjust height and width of right divs.  Also limit the .play div to only contain about 60% of width.
--
play.tmpl.htm

Add additional divs for instructions.  Have defintion <li> objects have the $index number on display.